reverse order of replication
Very large files that are still being uploaded when the repl script
kicks will show up as zero byte files. The script will attempt to
replicate them while they are still be uploaded. The replication will
fail during the checksum because the file's checksum has changed. To
give these ongoing uploads a little more time to finish, the order of
the replications will now start with the largest cohort first.