CORE-530 Change order of repl processing
Two changes were made.
1) The replication is performed so that the smaller files are replicated
before larger one.
2) Due to the fact that very large files may still be being uploaded
when the repl script determines that they need to be uploaded. These
files will appear as zero byte files. To give them a chance to finish
uploading before replication, the 0-31 MiB group is processed next to
last.
3) Since the largest files cause the replication throughput to drop off,
the the 480+ MiB group is processed last so that the throughput drop
effects the fewest number of files.