repl-report fixed bug in age to iRODS timestamp conv
I fixed a bug in conversation of age in days to an iRODS timestamp where
ages that converted to posix timestamps less than 1 billion had only one
leading zero in the generated iRODS timestamp. Now the generated iRODS
timestamps are padded on the left with "0"s to ensure a length of 11
characters.