repl: corrected resc child mapping
The generation is now done starting with the root resources instead of the leaf resources, because
r_resc_main.resc_children is not a valid way to determine if a resource has children. It is only
guaranteed to be populated if there is a context attached to the parent-child relationship.